+++ src/OctoshiftCLI.Tests/ado2gh/Commands/AddTeamToRepoCommandTests.cs
@@ -3,6 +3,7 @@ using System.Threading.Tasks;
 using Moq;
 using OctoshiftCLI.AdoToGithub;
 using OctoshiftCLI.AdoToGithub.Commands;
+using OctoshiftCLI.AdoToGithub.Handlers;
 using Xunit;
 
 namespace OctoshiftCLI.Tests.AdoToGithub.Commands
@@ -13,7 +14,7 @@ namespace OctoshiftCLI.Tests.AdoToGithub.Commands
         private readonly Mock<GithubApiFactory> _mockGithubApiFactory = TestHelpers.CreateMock<GithubApiFactory>();
         private readonly Mock<OctoLogger> _mockOctoLogger = TestHelpers.CreateMock<OctoLogger>();
 
-        private readonly AddTeamToRepoCommand _command;
+        private readonly AddTeamToRepoCommandHandler _command;
 
         private readonly string GITHUB_ORG = "foo-org";
         private readonly string GITHUB_REPO = "foo-repo";
@@ -21,22 +22,23 @@ namespace OctoshiftCLI.Tests.AdoToGithub.Commands
 
         public AddTeamToRepoCommandTests()
         {
-            _command = new AddTeamToRepoCommand(_mockOctoLogger.Object, _mockGithubApiFactory.Object);
+            _command = new AddTeamToRepoCommandHandler(_mockOctoLogger.Object, _mockGithubApiFactory.Object);
         }
 
         [Fact]
         public void Should_Have_Options()
         {
-            Assert.NotNull(_command);
-            Assert.Equal("add-team-to-repo", _command.Name);
-            Assert.Equal(6, _command.Options.Count);
+            var command = new AddTeamToRepoCommand(_mockOctoLogger.Object, _mockGithubApiFactory.Object);
+            Assert.NotNull(command);
+            Assert.Equal("add-team-to-repo", command.Name);
+            Assert.Equal(6, command.Options.Count);
 
-            TestHelpers.VerifyCommandOption(_command.Options, "github-org", true);
-            TestHelpers.VerifyCommandOption(_command.Options, "github-repo", true);
-            TestHelpers.VerifyCommandOption(_command.Options, "team", true);
-            TestHelpers.VerifyCommandOption(_command.Options, "role", true);
-            TestHelpers.VerifyCommandOption(_command.Options, "github-pat", false);
-            TestHelpers.VerifyCommandOption(_command.Options, "verbose", false);
+            TestHelpers.VerifyCommandOption(command.Options, "github-org", true);
+            TestHelpers.VerifyCommandOption(command.Options, "github-repo", true);
+            TestHelpers.VerifyCommandOption(command.Options, "team", true);
+            TestHelpers.VerifyCommandOption(command.Options, "role", true);
+            TestHelpers.VerifyCommandOption(command.Options, "github-pat", false);
+            TestHelpers.VerifyCommandOption(command.Options, "verbose", false);
         }
 
         [Fact]
@@ -88,7 +90,8 @@ namespace OctoshiftCLI.Tests.AdoToGithub.Commands
             _mockGithubApiFactory.Setup(m => m.Create(It.IsAny<string>(), It.IsAny<string>())).Returns(_mockGithubApi.Object);
 
             var root = new RootCommand();
-            root.AddCommand(_command);
+            var command = new AddTeamToRepoCommand(_mockOctoLogger.Object, _mockGithubApiFactory.Object);
+            root.AddCommand(command);
             var args = new string[] { "add-team-to-repo", "--github-org", GITHUB_ORG, "--github-repo", GITHUB_REPO, "--team", TEAM, "--role", role };
             await root.InvokeAsync(args);
