+++ src/ado2gh/Commands/DisableRepoCommand.cs
@@ -1,25 +1,18 @@
 ﻿using System;
 using System.CommandLine;
 using System.CommandLine.NamingConventionBinder;
-using System.Linq;
-using System.Threading.Tasks;
+using OctoshiftCLI.AdoToGithub.Handlers;
 
 namespace OctoshiftCLI.AdoToGithub.Commands
 {
     public class DisableRepoCommand : Command
     {
-        private readonly OctoLogger _log;
-        private readonly AdoApiFactory _adoApiFactory;
-
         public DisableRepoCommand(OctoLogger log, AdoApiFactory adoApiFactory) : base(
             name: "disable-ado-repo",
             description: "Disables the repo in Azure DevOps. This makes the repo non-readable for all." +
                          Environment.NewLine +
                          "Note: Expects ADO_PAT env variable or --ado-pat option to be set.")
         {
-            _log = log;
-            _adoApiFactory = adoApiFactory;
-
             var adoOrg = new Option<string>("--ado-org")
             {
                 IsRequired = true
@@ -47,39 +40,8 @@ namespace OctoshiftCLI.AdoToGithub.Commands
             AddOption(adoPat);
             AddOption(verbose);
 
-            Handler = CommandHandler.Create<DisableRepoCommandArgs>(Invoke);
-        }
-
-        public async Task Invoke(DisableRepoCommandArgs args)
-        {
-            if (args is null)
-            {
-                throw new ArgumentNullException(nameof(args));
-            }
-
-            _log.Verbose = args.Verbose;
-
-            _log.LogInformation("Disabling repo...");
-            _log.LogInformation($"ADO ORG: {args.AdoOrg}");
-            _log.LogInformation($"ADO TEAM PROJECT: {args.AdoTeamProject}");
-            _log.LogInformation($"ADO REPO: {args.AdoRepo}");
-            if (args.AdoPat is not null)
-            {
-                _log.LogInformation("ADO PAT: ***");
-            }
-
-            var ado = _adoApiFactory.Create(args.AdoPat);
-
-            var allRepos = await ado.GetRepos(args.AdoOrg, args.AdoTeamProject);
-            if (allRepos.Any(r => r.Name == args.AdoRepo && r.IsDisabled))
-            {
-                _log.LogSuccess($"Repo '{args.AdoOrg}/{args.AdoTeamProject}/{args.AdoRepo}' is already disabled - No action will be performed");
-                return;
-            }
-            var repoId = allRepos.First(r => r.Name == args.AdoRepo).Id;
-            await ado.DisableRepo(args.AdoOrg, args.AdoTeamProject, repoId);
-
-            _log.LogSuccess("Repo successfully disabled");
+            var handler = new DisableRepoCommandHandler(log, adoApiFactory);
+            Handler = CommandHandler.Create<DisableRepoCommandArgs>(handler.Invoke);
         }
     }
