+++ src/ado2gh/Commands/ShareServiceConnectionCommand.cs
@@ -1,24 +1,18 @@
 ﻿using System;
 using System.CommandLine;
 using System.CommandLine.NamingConventionBinder;
-using System.Threading.Tasks;
+using OctoshiftCLI.AdoToGithub.Handlers;
 
 namespace OctoshiftCLI.AdoToGithub.Commands
 {
     public class ShareServiceConnectionCommand : Command
     {
-        private readonly OctoLogger _log;
-        private readonly AdoApiFactory _adoApiFactory;
-
         public ShareServiceConnectionCommand(OctoLogger log, AdoApiFactory adoApiFactory) : base(
             name: "share-service-connection",
             description: "Makes an existing GitHub Pipelines App service connection available in another team project. This is required before you can rewire pipelines." +
                          Environment.NewLine +
                          "Note: Expects ADO_PAT env variable or --ado-pat option to be set.")
         {
-            _log = log;
-            _adoApiFactory = adoApiFactory;
-
             var adoOrg = new Option<string>("--ado-org")
             {
                 IsRequired = true
@@ -46,40 +40,8 @@ namespace OctoshiftCLI.AdoToGithub.Commands
             AddOption(adoPat);
             AddOption(verbose);
 
-            Handler = CommandHandler.Create<ShareServiceConnectionCommandArgs>(Invoke);
-        }
-
-        public async Task Invoke(ShareServiceConnectionCommandArgs args)
-        {
-            if (args is null)
-            {
-                throw new ArgumentNullException(nameof(args));
-            }
-
-            _log.Verbose = args.Verbose;
-
-            _log.LogInformation("Sharing Service Connection...");
-            _log.LogInformation($"ADO ORG: {args.AdoOrg}");
-            _log.LogInformation($"ADO TEAM PROJECT: {args.AdoTeamProject}");
-            _log.LogInformation($"SERVICE CONNECTION ID: {args.ServiceConnectionId}");
-            if (args.AdoPat is not null)
-            {
-                _log.LogInformation("ADO PAT: ***");
-            }
-
-            var ado = _adoApiFactory.Create(args.AdoPat);
-
-            var adoTeamProjectId = await ado.GetTeamProjectId(args.AdoOrg, args.AdoTeamProject);
-
-            if (await ado.ContainsServiceConnection(args.AdoOrg, args.AdoTeamProject, args.ServiceConnectionId))
-            {
-                _log.LogInformation("Service connection already shared with team project");
-                return;
-            }
-
-            await ado.ShareServiceConnection(args.AdoOrg, args.AdoTeamProject, adoTeamProjectId, args.ServiceConnectionId);
-
-            _log.LogSuccess("Successfully shared service connection");
+            var handler = new ShareServiceConnectionCommandHandler(log, adoApiFactory);
+            Handler = CommandHandler.Create<ShareServiceConnectionCommandArgs>(handler.Invoke);
         }
     }
